moved wallet/util.py -> network/liquid/util.py moved wallet/utils.py -> wallet/seed.py
