+++ tests/test_encryption.py
@@ -1,6 +1,6 @@
 import sha3
 import nacl.secret
-from x21e8.wallet.utils import create_and_save_seed
+from x21e8.wallet.seed import create_and_save_seed
 from x21e8.encryption import encrypt_str, encrypt_dict, encrypt_bytes, decrypt_2_str, decrypt_2_dict, decrypt_2_bytes
 
 try:
