+++ x21e8/network/liquid/liquid_network_definitions.py
@@ -1,0 +1,9 @@
+# define BIP32_PATH_MAX_LEN 255
+
+# /** Indicate that we want to derive a private key in `bip32_key_from_parent` */
+# /** Indicate that we want to derive a public key in `bip32_key_from_parent` */
+# /** Indicate that we want to skip hash calculation when deriving a key in `bip32_key_from_parent` */
+# /** Indicate that we want the pub tweak to be added to the calculation when deriving a key in `bip32_key_from_parent` */
+# /** Only used with elements */
+# /** Allow a wildcard ``*`` or ``*'``/``*h`` in path string expressions */
+# /** Do not allow a leading ``m``/``M`` or ``/`` in path string expressions */
